+++ packages/assemble-release-plan/src/increment.ts
@@ -1,6 +1,6 @@
 import semverInc from "semver/functions/inc";
 import { InternalRelease, PreInfo } from "./types";
-import { InternalError } from "@changesets/errors";
+import { mapGetOrThrowInternal } from "./utils";
 
 export function incrementVersion(
   release: InternalRelease,
@@ -12,12 +12,11 @@ export function incrementVersion(
 
   let version = semverInc(release.oldVersion, release.type)!;
   if (preInfo !== undefined && preInfo.state.mode !== "exit") {
-    let preVersion = preInfo.preVersions.get(release.name);
-    if (preVersion === undefined) {
-      throw new InternalError(
-        `preVersion for ${release.name} does not exist when preState is defined`
-      );
-    }
+    let preVersion = mapGetOrThrowInternal(
+      preInfo.preVersions,
+      release.name,
+      `preVersion for ${release.name} does not exist when preState is defined`
+    );
     // why are we adding this ourselves rather than passing 'pre' + versionType to semver.inc?
     // because semver.inc with prereleases is confusing and this seems easier
     version += `-${preInfo.state.tag}.${preVersion}`;