+++ packages/changelog-github/src/index.test.ts
@@ -118,6 +118,112 @@ describe.each(["author", "user"])(
   }
 );
 
+it("linkifies bare issue references", async () => {
+  expect(
+    await getReleaseLine(...getChangeset("fixes #1234 and #5678", data.commit))
+  ).toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        fixes [#1234](https://github.com/emotion-js/emotion/issues/1234) and [#5678](https://github.com/emotion-js/emotion/issues/5678)"
+  `);
+});
+
+it("does not double-linkify existing markdown links", async () => {
+  expect(
+    await getReleaseLine(
+      ...getChangeset(
+        "see [#1234](https://github.com/emotion-js/emotion/issues/1234)",
+        data.commit
+      )
+    )
+  ).toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        see [#1234](https://github.com/emotion-js/emotion/issues/1234)"
+  `);
+});
+
+it("does not linkify issue-like refs inside link text", async () => {
+  expect(
+    await getReleaseLine(
+      ...getChangeset("see [fix for #99](https://example.com)", data.commit)
+    )
+  ).toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        see [fix for #99](https://example.com)"
+  `);
+});
+
+it("does not linkify when preceded by a word character", async () => {
+  expect(await getReleaseLine(...getChangeset("foo#123", data.commit)))
+    .toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        foo#123"
+  `);
+});
+
+it("does not linkify #0", async () => {
+  expect(await getReleaseLine(...getChangeset("see #0", data.commit)))
+    .toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        see #0"
+  `);
+});
+
+it("linkifies issue ref at the start of a line", async () => {
+  expect(await getReleaseLine(...getChangeset("#42 was fixed", data.commit)))
+    .toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        [#42](https://github.com/emotion-js/emotion/issues/42) was fixed"
+  `);
+});
+
+it("linkifies issue ref after punctuation", async () => {
+  expect(await getReleaseLine(...getChangeset("fixed (#99)", data.commit)))
+    .toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        fixed ([#99](https://github.com/emotion-js/emotion/issues/99))"
+  `);
+});
+
+it("handles mixed linked and bare refs", async () => {
+  expect(
+    await getReleaseLine(
+      ...getChangeset(
+        "fixes [#1](https://github.com/emotion-js/emotion/issues/1) and #2",
+        data.commit
+      )
+    )
+  ).toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        fixes [#1](https://github.com/emotion-js/emotion/issues/1) and [#2](https://github.com/emotion-js/emotion/issues/2)"
+  `);
+});
+
+it("linkifies issue ref followed by a dot", async () => {
+  expect(await getReleaseLine(...getChangeset("this fixes #42.", data.commit)))
+    .toMatchInlineSnapshot(`
+    "
+
+    - [#1613](https://github.com/emotion-js/emotion/pull/1613) [\`a085003\`](https://github.com/emotion-js/emotion/commit/a085003) Thanks [@Andarist](https://github.com/Andarist)! - something
+        this fixes [#42](https://github.com/emotion-js/emotion/issues/42)."
+  `);
+});
+
 it("with multiple authors", async () => {
   expect(
     await getReleaseLine(