+++ packages/get-github-info/src/index.ts
@@ -1,6 +1,14 @@
-// @ts-ignore
 import fetch from "node-fetch";
 import DataLoader from "dataloader";
+
+function readEnv() {
+  const GITHUB_GRAPHQL_URL =
+    process.env.GITHUB_GRAPHQL_URL || "https://api.github.com/graphql";
+  const GITHUB_SERVER_URL =
+    process.env.GITHUB_SERVER_URL || "https://github.com";
+  const GITHUB_TOKEN = process.env.GITHUB_TOKEN;
+  return { GITHUB_GRAPHQL_URL, GITHUB_SERVER_URL, GITHUB_TOKEN };
+}
 
 const validRepoNameRegex = /^[\w.-]+\/[\w.-]+$/;
 
@@ -79,9 +87,10 @@ function makeQuery(repos: ReposWithCommitsAndPRsToFetch) {
 // getReleaseLine will be called a large number of times but it'll be called at the same time
 // so instead of doing a bunch of network requests, we can do a single one.
 const GHDataLoader = new DataLoader(async (requests: RequestData[]) => {
-  if (!process.env.GITHUB_TOKEN) {
+  const { GITHUB_GRAPHQL_URL, GITHUB_SERVER_URL, GITHUB_TOKEN } = readEnv();
+  if (!GITHUB_TOKEN) {
     throw new Error(
-      `Please create a GitHub personal access token at https://github.com/settings/tokens/new?scopes=read:user,repo:status&description=changesets-${new Date()
+      `Please create a GitHub personal access token at ${GITHUB_SERVER_URL}/settings/tokens/new?scopes=read:user,repo:status&description=changesets-${new Date()
         .toISOString()
         .substring(
           0,
@@ -97,17 +106,31 @@ const GHDataLoader = new DataLoader(async (requests: RequestData[]) => {
     repos[repo].push(data);
   });
 
-  const data = await fetch("https://api.github.com/graphql", {
-    method: "POST",
-    headers: {
-      Authorization: `Token ${process.env.GITHUB_TOKEN}`,
-    },
-    body: JSON.stringify({ query: makeQuery(repos) }),
-  }).then((x: any) => x.json());
+  let fetchResponse;
+  try {
+    fetchResponse = await fetch(GITHUB_GRAPHQL_URL, {
+      method: "POST",
+      headers: {
+        Authorization: `Token ${GITHUB_TOKEN}`,
+      },
+      body: JSON.stringify({ query: makeQuery(repos) }),
+    });
+  } catch (e: any) {
+    throw new Error(
+      `An error occurred when fetching data from GitHub\n${e.message}`
+    );
+  }
+
+  let data;
+  try {
+    data = await fetchResponse.json();
+  } catch (e: any) {
+    throw new Error(`Failed to parse data from GitHub\n${e.message}`);
+  }
 
   if (data.errors) {
     throw new Error(
-      `An error occurred when fetching data from GitHub\n${JSON.stringify(
+      `Fetched data from GitHub returned errors\n${JSON.stringify(
         data.errors,
         null,
         2
@@ -118,9 +141,7 @@ const GHDataLoader = new DataLoader(async (requests: RequestData[]) => {
   // this is mainly for the case where there's an authentication problem
   if (!data.data) {
     throw new Error(
-      `An error occurred when fetching data from GitHub\n${JSON.stringify(
-        data
-      )}`
+      `Fetched data from GitHub has missing data\n${JSON.stringify(data)}`
     );
   }
 
@@ -262,7 +283,7 @@ export async function getInfoFromPullRequest(request: {
       commit: commit
         ? `[\`${commit.abbreviatedOid.slice(0, 7)}\`](${commit.commitUrl})`
         : null,
-      pull: `[#${request.pull}](https://github.com/${request.repo}/pull/${request.pull})`,
+      pull: `[#${request.pull}](${data.url})`,
       user: user ? `[@${user.login}](${user.url})` : null,
     },
   };